marketplace/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
grpcio-tools ~= 1.30
mypy-protobuf ~= 2.4
pytest ~= 5.4
